+++ CHANGELOG.md
@@ -4,10 +4,12 @@
   * Retrait du menu en mode jeu
     * Les grilles sont toujours centrées (hors scrolling)
   * Capture d'écran si SCREEN_CAPTURE défini
+  * Sortie du jeu => "start" est grisé
 
 * Corrections
     * BUG : Rotation => pos = {0,0}
     * BUG : en mode DEBUG, affichage des statss même lorsque la 'grid' est vide
+    * BUG : _onStep => affichage en dehors du viewPort possible
 
 #### v0.0.1 - beta 4
 * 20 mars 2025